test: avoid node confusion over entrypoint
Running `node .` in `test/app/` causes it to look for `test/app.js`,
find it, and run it... I guess it makes sense, kindof, though running .
and having it run something in .. is somewhat surprising. Anyhow,
renaming app.js allows node to not find it, and instead run ./index.js
(the package.json main) as intended.