test: align app exit status with strong-supervisor
test app was converting signals to exit codes, which was fine but when I
run the same tests with sl-run, it was exiting by signal not
with a status code. app now does as sl-run does, and signal comparison
is now done in terms of signal strings, not exit codes.