Upgrade strong-service-install to 1.0.0
Add test for sl-pm path in generated init script to ensure the
installer is generating correct jobs.

New version includes bug fixes for environment variables and execPath
handling.

New version also uses the "upstart" option to indicate the version of
Upstart to generate the job file for instead of the "version".