debug: fix debug output suggesting process leaks
On success, when the supervisor exits, any of its children are updated
in case they weren't previously marked stopped. If they were previously
marked stopped, the usual case, for every ServiceProcess, the following
message would be printed (in debug mode). Under some conditions, this
would be true, but if the process was found, and the stopTime was
already set, this message was wrong and misleading.

    strong-pm:server Got exit event for an unknown process: {"cmd":"exit",...

Fix strongloop-internal/scrum-nodeops#226