Support sl-run spawn on Windows
We were injecting all our dependencies .bin dirs into our path, and
expecting spawn would find them. On Windows, node scripts aren't
executable, so it doesn't matter if they are in the PATH. So, we
tokenize the start command, naively, but good enough for now
(configurability of the start command is going away ASAP), and spawn
node to execute the supervisor script.