feature: pmctl log-dump command to dump logs
Running 'pmctl log-dump' will read the contents of the buffer and then
purge the buffer as an atomic operation as a quick'n'dirty substitute
for proper pagination or state tracking. The log-dump command also
supports a --follow mode which is essentially repeats the command every
second until stopped, giving an approximation of live log streaming.

Also extend helper-async to accomodate cleaner tests.