test: helper-async failon was ignoring pattern
So failure for any reason was considered OK, even if the pattern didn't
match. This masked bugs such as calling 'stop' with not service arg when
it was expected to fail because it was already stopped, not because the
syntax is invalid.

Asserting on the pattern may break existing tests that seemed to be
passing.