driver: implement changing cluster size
Previously, a change in the instance.cpus value caused the start command
to be updated, but this didn't take effect under most scenarios.

On initial start of an instance from the saved meta, the supervisor was
being started with the default value for size, not the instance.cpus.

When a supervisor was already running, only the start command was
changing, but that has no effect on a supervisor that is already
started.

With this change, the expected size is included in the instance meta, so
effects the initial start of the supervisor. Also, when the size
changes, a set-size request is sent to the currently running supervisor,
if any, so that it can change its size dynamically.