refactor: remove Driver#_containerById interface
The only thing it was used for was to get the .current member of the
object that is returned, so it was a private method, called publicly,
and only used to get at implementation details of the opaque object
that was returned.

So instead, define a Driver#instanceById interface that returns the
object that is used directly, without depending on any implementation
details. Now drivers aren't forced to have a ".current" member on their
internal tracking objects.