test: update expectations for persisted set-size
My only guess as to why it was waiting for a condition that will
logically never happen was because the size was not being persisted
previously and the cluster would revert back to the default state.

Either that or it was depending on the `pmctl` running just as the
process count was coming up, but not before it was fully restored.