Default base dir in the home directory
Putting it in CWD makes sense if you can run multiple copies of pm
simultaneously, but you cannot, they all listen on port 8701. If you
want to put them on different ports, you can, but it requires passing
command line options, in which case, you can use --base if you want
different ports and different base dirs.

This should help unify the various ways of starting the pm, as well
avoiding the problem where arc assumes that the .strong-pm in it's CWD
belongs to it, and deletes it on exit, causing pm to crash.