server: skip 'status' after 'started' notification
All the information we need is now in the 'started' message, so there
is no need for querying with a 'status' message to get additional
details. This saves a round trip and a potential source of race
conditions due to subsequent notifications being processed before the
instance has officially "started".