docker: make driver startup more reliable
The debian image is actually a subset of the node image, so pulling
them in parallel doesn't actually do anything except cause docker to
return messages saying the image is already being pulled.

This way, when the debian image gets pulled it is almost a no-op since
all it has to do is tag one of the layers that was just downloaded.