docker: remove non-websocket ctl channel support
The supervisor initiated websocket control channel allows removal of
the steps in the startup to get port mappings and then initiate a plain
TCP based control connection (and reconnection logic). It also allows
the faking of a 'started' message to be refactored to only be done when
actually resuming an intance when its container is still running from a
previous deployment.

The 'started' message synthesis is further simplified because the
response to a 'status' request now includes all of the information
needed to create it.