service-manager: allow application port to be set
Normally, pm sets PORT to a value guaranteed to be different for each
app, so that multiple apps can be run without their ports colliding.
However, some apps may require being run on a specific port. This is
supported by doing a

  slc ctl set-env <SVC> PORT=<X>

Where X is the port that should be used. If this port is already used,
then the app will immediately crash, but that's a configuration problem
under the user's control.