test: update docker tests to use official node
When these tests were written there was no official node image. Now the
nodesource image no longer provides the image that was originally used.

Simple fix is to just tweak the Dockerfile to use a different image, and
adjust for it.

Note that this is mostly to unblock upstream builds.